Update Set-NewLDAPSCertificate.ps1
Updated script to be used in auto renewing LDAPS expiring cert

+++ Set-NewLDAPSCertificate.ps1
@@ -3,7 +3,7 @@
 # https://github.com/tobor88/PowerShell/blob/master/Hide-PowerShellScriptPassword.ps1
 $KeyPassword = "LDAPS-S3cr3t-Pa55w0rd" # I have a script at the above link you can use to encrypt this value so it does not show in clear text
 $SecurePassword = ConvertTo-SecureString -String $KeyPassword -Force –AsPlainText
-$CertPath = "$env:USERPROFILE\Downloads\LDAPS.pfx"
+#$CertPath = "$env:USERPROFILE\Downloads\LDAPS.pfx"
 $LDAPSTemplateName = "LDAP over SSL"
 $ServiceNames = "NTDS",((Get-CimInstance -ClassName Win32_Service -Filter 'Name LIKE "%ADAM%"').Name)
 # NTDS is the default LDAP service. 
@@ -24,11 +24,11 @@ If (($LDAPSCert -Contains $ExpiringCert) -and ($Null -ne $LDAPSCert[1])) {
 ElseIf ($LDAPSCert -eq $ExpiringCert) {
 
     Write-Output "[*] Renewing LDAPS certificate with the same keys"
-    Start-Process -WorkingDirectory "C:\Windows\System32" -FilePath certreq.exe -ArgumentList @('-Enroll', '-machine', '-q', '-cert', $LDAPSCert.SerialNumber, 'Renew', 'ReuseKeys')
+    Start-Process -WorkingDirectory "C:\Windows\System32" -FilePath certreq.exe -ArgumentList @('-Enroll', '-machine', '-q', '-cert', $LDAPSCert.SerialNumber, 'Renew', 'ReuseKeys') -Wait
 
     $LDAPSCert = Get-ChildItem -Path Cert:\LocalMachine\My | Where-Object -FilterScript { $_.Extensions | Where-Object -FilterScript { ($_.Oid.FriendlyName -eq "Certificate Template Information") -and ($_.Format(0) -Match $LDAPSTemplateName) }}
-    $ExpiringCert = Get-ChildItem -Path Cert:\LocalMachine\My -ExpiringInDays 30 | Where-Object -FilterScript { $_.Extensions | Where-Object -FilterScript { ($_.Oid.FriendlyName -eq "Certificate Template Information") -and ($_.Format(0) -Match $LDAPSTemplateName) }}
-    If ($LDAPSCert -ne $ExpiringCert) {
+    $ExpiringCert = Get-ChildItem -Path Cert:\LocalMachine\My -ExpiringInDays 90 | Where-Object -FilterScript { $_.Extensions | Where-Object -FilterScript { ($_.Oid.FriendlyName -eq "Certificate Template Information") -and ($_.Format(0) -Match $LDAPSTemplateName) }}
+    If (($LDAPSCert -ne $ExpiringCert) -and ($Null -ne $ExpiringCert)) {
 
         Write-Output "[*] Deleting the old certificate from the LocalMachine Certificate store. New certificate has a different thumbprint"
         Get-ChildItem -Path "Cert:\LocalMachine\My\$($ExpiringCert.Thubprint)" | Remove-Item -Force
@@ -69,3 +69,4 @@ ForEach ($ServiceName in $ServiceNames) {
     }  # End If
 
 }  # End ForEach
+
